requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Pillow==8.2.0
Werkzeug==2.0.1
